+++ .storybook/preview.js
@@ -50,6 +50,10 @@ import '../source/patterns/elements/collapse/collapse.js';
 
 // Import collection behaviors
 import '../source/patterns/collections/accordion/accordion.js';
+import '../source/patterns/collections/menu-primary/menu-primary.js';
+
+// Import layout behaviors
+import '../source/patterns/layouts/header/header.js';
 
 // Import base pattern behaviors
 import '../source/patterns/base/animations/animations.js';
@@ -192,7 +196,7 @@ const preview = {
         includeName: true,
       },
     },
-    viewport: { viewports: customViewports },
+    viewport: { options: customViewports },
   },
   controls: {
     matchers: {